Dominican Republic (Diputados) add term-2010 to instructions
Both the official site and wikipedia seem to be showing only the
data for the 2016 term, so at the moment morph scraper is mixing
the data already available from term 2010 with the new incoming
data for term 2016. This commit adds the manual file with the data
from term-2010 as a manual source to the instructions.json file.