Use the bundler github shortcut in Gemfile
This overrides the `git_source` definition for :github to force it to
use https:// and uses the :github shortcut to point to the gem's
repos when they are not in Rubygems.

By default the `:github` shortcut in Bundler uses the git:// protocol,
which is insecure. This leaves us vulnerable to a man-in-the-middle
attack which could change the code as we check it out and potentially
replace it with something malicious.

See http://bundler.io/git.html#security for more details.