Northern Mariana Islands (House): unreconcile Lorenzo I. De Leon Guerrero
Q3828608 is the wrong Lorenzo De Leon Guerrero: this one died before the
current term started. The current member might be related to him, but
this isn't the correct ID.

However, as this is the only person we have here, removing this line
entirely will cause the build to fail with an empty file. So, for now,
as a nasty workaround, explicitly reconcile to Q0.