Add UUIDs for Areas if remapped
Similarly to how we did for Groups, if there's a UUID mapping file for
the areas for a given source, switch those in.

There's some duplication here that could porentially be removed, but
it's not really clear what the abstraction could/should be, so let's
hold off on doing anything with that until we've actually added UUIDs
for quite a few legislatures, in case there are any odd gotchas we'll
need to handle differently, where premature abstraction could be more of
a hindrance than a help.