Don't display unmatched reconciliation if 'report_missing' set
This presumably existed at some point in the past, as we have a flag for
this set in a variety of instructions files, but it has disappeared.
Bring it back.

This is usually for the case where a Wikidata source includes lots of
details for former members who we don't know anything about, and we
don't need to clutter up the build output with lots of "unmatched"
warnings, especially if we've actually matched every member.