add a PROJECT Pathname
This is the root of the repo itself, and can be used to navigate into
areas outside the legislature we’re working with.

We should possibly call `realpath` on this at this point rather than
requiring (some) callers to do so, but it’s not clear-cut, so leave it
as a relative path for now.