Update README
The section on how to add a new country/legislature is now very out of
date, and misleading. This is such a rare occurence now that we should
really remove it, rather than updating it.

Similarly, the information on refreshing `countries.json` is no longer
relevant.

We can also now rebuild individual sources, so should mention that too.